Apparently fix a bug with EMNIST-like datasets (R)
The EMNIST dataset (and other like it) are pretty easy to download. The
code was raising an error whenever it wasn't able to find the root for
the given dataset on the current cluster, which isn't correct! We'd like
to instead just download the dataset to SCRATCH/data, and then perhaps
copy it into $SLURM_TMPDIR or something similar.

Signed-off-by: Fabrice Normandin <fabrice.normandin@gmail.com>